requirements.txt
# =============================================================================
# SALESFORCE ANALYTICS EMBEDDING TEMPLATE - PYTHON REQUIREMENTS
# =============================================================================

# Core Flask application
Flask==3.0.0
Flask-CORS==4.0.0

# Environment and configuration management
python-dotenv==1.0.0
environs==10.0.0

# HTTP requests and OAuth
requests==2.31.0
requests-oauthlib==1.3.1
oauthlib==3.2.2

# Security and authentication
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.0

# Session management
Flask-Session==0.5.0
redis==5.0.1

# Database support (optional)
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9  # PostgreSQL support
pymongo==4.6.0  # MongoDB support

# Caching
Flask-Caching==2.1.0

# Rate limiting
Flask-Limiter==3.5.0

# JSON handling and validation
jsonschema==4.20.0
marshmallow==3.20.1

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Error tracking (optional integrations)
sentry-sdk[flask]==1.38.0

# Development and testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API documentation
flasgger==0.9.7.1

# Excel/CSV export capabilities
pandas==2.1.4
openpyxl==3.1.2

# Template rendering enhancements
Jinja2==3.1.2
MarkupSafe==2.1.3

# Health checks and utilities
Flask-HealthCheck==1.0.0

# Production server
gunicorn==21.2.0

# SSL/TLS support
pyOpenSSL==23.3.0

# Veeva CRM integration utilities
xmltodict==0.13.0  # XML parsing for Veeva APIs
defusedxml==0.7.1  # Secure XML parsing

# Industry-specific utilities
pycountry==22.3.13  # Country/region data for pharma
babel==2.13.1  # Internationalization

# Data processing and analytics
numpy==1.25.2
scipy==1.11.4

# Configuration validation
cerberus==1.3.5

# HTTP server enhancements
werkzeug==3.0.1

# Template utilities
click==8.1.7

# =============================================================================
# OPTIONAL DEPENDENCIES FOR SPECIFIC FEATURES
# =============================================================================

# Uncomment as needed based on your deployment and feature requirements:

# AWS deployment
# boto3==1.34.0
# botocore==1.34.0

# Azure deployment
# azure-storage-blob==12.19.0
# azure-identity==1.15.0

# Google Cloud deployment
# google-cloud-storage==2.10.0
# google-auth==2.25.0

# Advanced analytics
# scikit-learn==1.3.2
# plotly==5.17.0

# Real-time features
# websockets==12.0
# Flask-SocketIO==5.3.6

# Message queues
# celery==5.3.4
# redis==5.0.1

# Email notifications
# Flask-Mail==0.9.1

# File upload handling
# Pillow==10.1.0

# Advanced caching
# pylibmc==1.6.3  # Memcached support

# =============================================================================
# DEVELOPMENT DEPENDENCIES
# =============================================================================

# Code quality
autopep8==2.0.4
isort==5.12.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Load testing
locust==2.17.0

# =============================================================================
# SECURITY NOTES
# =============================================================================

# All versions are pinned for security and reproducibility
# Run 'pip-audit' to check for vulnerabilities
# Update versions regularly and test thoroughly

# For production, consider:
# - Using pip-tools for dependency management
# - Setting up vulnerability scanning
# - Using virtual environments
# - Regular security audits